Special-case Builder in relative-path validation
Since the Builder environment doesn't place code in "/code" some of the
assumptions made by the CLI need to be worked around. We do this on one side by
modifying the environment (CODECLIMATE_CODE), but in this case we have to ignore
MountedPath.code.container_path and hardcode what it is in Builder.

The actual solution is probably do inject a MountedPath object through the
Bridge, so the local and container-path properties can be specific to where the
class is being run. If we're unable to move to sub-commands and/or always
invoking CLI as a process, we might take on that refactor later.