Config::Engine#config is the entire config
Working with config.config was confusing, but there is good reason: we always
want to read and store config.checks, config.exclude_fingerprints, etc. Putting
the whole Hash as-is into the Config::Engine object is overall simpler.

Note: this is safe because of our existing assumption that this code works on
pre-validated configuration.