Cache the other_location check by path. (#582)
* Cache the other_location check by path.

This speeds up anything that makes use of other_locations extensively.
If one location is valid, all of them (with the same path) will be.
This took codeclimate-duplication on big_files from 603s to 108s.

More could probably be done to speed up the driver, but this was
definitely at the top of my profile.

* Change failing test to have consistent other_location data.

Either the generator will generate entirely valid locations or will
generate entirely invalid locations. The likelihood of generating a mix
of the two is miniscule (tho plausible, still GIGO).