More-thoroughly validate location
We started seeing some exceptions when generating a default fingerprint
for an issue with a malformed location. We already have a validation to
prevent processing default engine output, but it's not enough. The
reason is that we check if the fingerprint is excluded _before_ we
validate the issue, and we can't (necessarily) generate a fingerprint
for an invalid issue. We were already validating the location in this
  code path, but not thoroughly.

This change doesn't avoid an error, but it gets us a better error by
leveraging the existing, thorough validation logic.

The next step is to handle this case in the pre-validation usage of
fingerprint.