Avoid error when inferring fingerprint
If the location is invalid, we may not be able to infer a fingerprint.
We need to be able to infer a fingerprint before we validate it (unless
we want to make a bigger refactor) so this change acknowledges that
sometimes we may not be able to do that, in which case let's _not_ skip
the issue, and let the validators evaluate it later in the pipe.