Add partial scan support
Notable changes:

* Added partial scan mode
* Fixed a dir node reexpansion after child removal

Partial Scan mode
=================

This is a new way to construct a workspace for analysis. It's similar to
target file mode when one can specify files for analysis except it takes
into account `exclude_paths`.

This is useful in particular for editor integration where one would want
to get edited files automatically analyzed but not get reports for
excluded files. The results of a partial scan are always a subset of
a full scan.