Config#disable_plugins! and #auto_enable_plugins?
To backfill repositories with historical analysis, we need to run builds
on commits many years old. Such commits might have code and
configuration that expects the then-latest version of engine tooling
(e.g. rubocop). This causes the builds to fail.

Our own engines should be stable to this (or we will make it so), but we
can't in any way ensure recent plugins would run on older codebases.
Therefore, we'll not run any plugins in such builds.

This will knowingly create builds that are missing any issues or
analysis data provided by plugins. Because of this, they can't be used
to render results on site or as merge-base analyses in comparisons.
We're comfortable with this limitation.