Support "plugins" in JSON config
This is the new key for specifying plugin engines.

Since we'll shortly be ignoring legacy yaml entirely when JSON is
present, we also have a practical need to be able to configure
structure/duplication in this file (e.g. to set channel), so this will
respect configuration for those engines under the "plugins" key, even
though they are not semantically plugins.