Either YAML or JSON, never both
With plugins, we now support all core config fields in the new schema
(excpect prepare, which I'll address soon). To simplify the
understanding of the system, we're moving towards only loading one of
these files: right now, the newer-schema JSON is the pre-eminent one, so
we won't load YAML if the JSON exists.