When auto-enabling, don't write `config: FILENAME`
This is an older style of config that engines should not rely on
anymore: though some still handle it, others will crash.

When auto-enabling an engine based on the config file existing, the
explicit config should not be needed anyway: since it's the auto-enable
path, it should be one of the paths the engine will look for by default
anyway.