Convert old file config to newer config style
The [cc-yaml][] config has been doing the conversion in this direction
for a long time, so I believe pretty much all engines support the newer
config format now. We primarily keep the conversion to support legacy
customer configs, not legacy engines, AFAIK. If there are any lingering
engines, we should fix those instead of re-writing to an older config
style, since the old-style config is *not supported* and will crash on
some engines.

[cc-yaml]: https://github.com/codeclimate/codeclimate-yaml/blob/master/spec/cc/yaml/nodes/engine_config_spec.rb#L58-L70