Treat YAML config as an adapter, not a separate config
This is the first part of a larger effort to make config treatment more
consistent, and to address lingering incompatibilities and edge cases in
Quality Model.

The core idea is that the user-provided config will go through a series
of transformations, but that instead of the transformations being
largely parallel as they are today (where YAML & JSON are implemented
completely in parallel), they will instead be a *series* of
transformations that build on each other. In this particular PR, the way
that works is that legacy YAML configs have their raw contents modified
to match newer schemas, and then go through the new JSON config logic.