Config: make defaults an adapter
Another step towards config handling being a single sequence of steps.
Defaults can be applied on top of whatever was already sourced from an
existing config file.

There are some nice simplifications I did because they were easy: e.g.
`duplication` no longer requires `languages` in the config, so we don't
need to set it.

This also helps fix some confusing behavior we've seen internally with
trying to override channels/enabled status for QM engines, and paves the
way for similar support for per-engine exclude_patterns, which we'll
restore soon. It's not *perfect* that we put these engines under
`plugins`, since they're not "plugins", but for now I think it's fine
that this is an implementation detail: customers won't tweak these bits
as much, and the technical benefits of treating these engines the same
as others is pretty large.