Prefer YAML.safe_load to SafeYaml
We've been mixing-and-matching these two approaches for a while. This
removes the final usage of `SafeYaml` and removes the gem.

`SafeYaml` is a 3rd party gem unassociated with the core Ruby team, and
it hasn't been touched in years. `YAML.safe_load` is a core part of
`Psych` distributed with Ruby, so it's naturally going to get more
modern updates & attention that `SafeYaml` will not. I ran many of
`SafeYaml`'s own examples through `YAML.safe_load` and `YAML`'s behavior
is equally correct.