Fix config handling of `analyze -e`.
We had a regression in QM with respecting engine config when only
certain engines are selected for running with `-e` or `--engine`.

The change here makes it so that `-e` will not reset any
already-specified config for an engine, but will still add a new engine
with default config when the named engine wasn't already configured.