Remove non functioning check for upconvert behavior
This was intended to mean we wouldn't do the upconvert if a `version`
key was specified, but it didn't get updated when we changed the
`schema_version` key to `version`, so it hasn't actually been working
anway.

Rather than fix it to actually work, I'm removing it: it should be safe
to do the upconvert all the time, and the current validations are not
written to account for the distinction with a YAML file with a `version`
key or not. I wouldn't want to lead into weird cases where somebody
added `version: 2` to a legacy file without changing anything else, and
`validate-config` told them everything was fine without actually using
their config as expected.