Rewrite `Net::HTTP` calls to support `http_proxy`
This is one piece of support for HTTP proxies with the `fetch` directive
in `.codeclimate.yml`.

`Net::HTTP` supports an `http_proxy` environment variable. It's
convenient in some respects, but requires that `Net::HTTP` be directly
instantiated with `#new`, as opposed to one of the alternate
constructors like `.start` or `.get_response`.

For more, please see:
https://www.jethrocarr.com/2014/08/14/ruby-nethttp-proxies/