Move to ruby 2.5
The recent PRs around ruby version reminded me that 2.3 is actually very
near it's end-of-life. 2.3 will no longer be supported for security
patches after the end of this month, so as long as we're here this seems
like a good moment to upgrade to 2.5.

(Also 2.3 is old enough that it's not building easily on my system,
making it tricky to cut a release, and I'd rather upgrade ruby for the
CLI than figure out how to build an almost-unsupported old version of
Ruby on my machine.)